Make getCorrection take delta as argument
Using the functor objects complicated things and when passed as reference created bad side effects.
It is now the caller's respoinsibility to supply the proper delta every time `getCorrection` is called.
IT will often be the case that the same delta value will be used by multiple PIDControllers anyways so this also reduces the redundent calls to getDelta.